Update 7/9
Added a bunch of useful stuff. Notes, local variables necessary for
encryption, a container method to contain everything, a method to read
in input from a text/hex file, sub_bytes method, a method to print out
the state array, and a method to trim the prefix string from the
hex/bin methods. Also, a method to print a delimiter.